# openapi_backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClaimVision-API",
    "description": "API del proyecto ClaimVision, en constante actualizacion y mejoria",
    "version": "1.12.0"
  },
  "paths": {
...
    "/api/v1/aseguradora/crud/ajustadores": {
      "post": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Crear Ajustador",
        "operationId": "crear_ajustador_api_v1_aseguradora_crud_ajustadores_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AjustadorCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AjustadorResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Listar Ajustadores",
        "operationId": "listar_ajustadores_api_v1_aseguradora_crud_ajustadores_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "activo_para_servicio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Activo Para Servicio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AjustadorResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aseguradora/crud/ajustadores/{id}": {
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Obtener Ajustador",
        "operationId": "obtener_ajustador_api_v1_aseguradora_crud_ajustadores__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AjustadorResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Actualizar Ajustador",
        "operationId": "actualizar_ajustador_api_v1_aseguradora_crud_ajustadores__id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AjustadorUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AjustadorResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Eliminar Ajustador",
        "operationId": "eliminar_ajustador_api_v1_aseguradora_crud_ajustadores__id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aseguradora/crud/clientes": {
      "post": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Crear Cliente",
        "operationId": "crear_cliente_api_v1_aseguradora_crud_clientes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Listar Clientes",
        "operationId": "listar_clientes_api_v1_aseguradora_crud_clientes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ClienteResponseDTO_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aseguradora/crud/clientes/{id}": {
      "get": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Obtener Cliente",
        "operationId": "obtener_cliente_api_v1_aseguradora_crud_clientes__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "v1 · Aseguradora CRUD"
        ],
        "summary": "Actualizar Cliente",
        "operationId": "actualizar_cliente_api_v1_aseguradora_crud_clientes__id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AjustadorCreateDTO": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "telefono": {
            "type": "string",
            "title": "Telefono"
          },
          "password_temporal": {
            "type": "string",
            "title": "Password Temporal"
          },
          "cedula_profesional": {
            "type": "string",
            "title": "Cedula Profesional"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "email",
          "telefono",
          "password_temporal",
          "cedula_profesional"
        ],
        "title": "AjustadorCreateDTO"
      },
...
      "AjustadorResponseDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "usuario_id": {
            "type": "string",
            "title": "Usuario Id"
          },
          "cedula_profesional": {
            "type": "string",
            "title": "Cedula Profesional"
          },
          "geolocalizacion_actual": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Geolocalizacion Actual"
          },
          "activo_para_servicio": {
            "type": "boolean",
            "title": "Activo Para Servicio"
          },
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "usuario_id",
          "cedula_profesional",
          "geolocalizacion_actual",
          "activo_para_servicio",
          "version",
          "created_at",
          "updated_at",
          "deleted_at"
        ],
        "title": "AjustadorResponseDTO"
      },
      "AjustadorUpdateDTO": {
        "properties": {
          "cedula_profesional": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cedula Profesional"
          },
          "activo_para_servicio": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activo Para Servicio"
          }
        },
        "type": "object",
        "title": "AjustadorUpdateDTO"
      },
...
      "ClienteCreateDTO": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "telefono": {
            "type": "string",
            "title": "Telefono"
          },
          "password_temporal": {
            "type": "string",
            "title": "Password Temporal"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "email",
          "telefono",
          "password_temporal"
        ],
        "title": "ClienteCreateDTO"
      },
      "ClienteResponseDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "usuario_id": {
            "type": "string",
            "title": "Usuario Id"
          },
          "numero_poliza": {
            "type": "string",
            "title": "Numero Poliza"
          },
          "vigencia_poliza": {
            "type": "string",
            "format": "date",
            "title": "Vigencia Poliza"
          },
          "consentimiento_aviso_privacidad": {
            "type": "boolean",
            "title": "Consentimiento Aviso Privacidad"
          },
          "consentimiento_biometria": {
            "type": "boolean",
            "title": "Consentimiento Biometria"
          },
          "autoriza_transferencia_talleres": {
            "type": "boolean",
            "title": "Autoriza Transferencia Talleres"
          },
          "fecha_consentimiento": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Consentimiento"
          },
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          }
        },
        "type": "object",
        "required": [
          "id",
          "usuario_id",
          "numero_poliza",
          "vigencia_poliza",
          "consentimiento_aviso_privacidad",
          "consentimiento_biometria",
          "autoriza_transferencia_talleres",
          "fecha_consentimiento"
        ],
        "title": "ClienteResponseDTO"
      },
      "ClienteUpdateDTO": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "numero_poliza": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Poliza"
          },
          "vigencia_poliza": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vigencia Poliza"
          }
        },
        "type": "object",
        "title": "ClienteUpdateDTO"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_AjustadorResponseDTO_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AjustadorResponseDTO"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[AjustadorResponseDTO]"
      },
...
      "Page_ClienteResponseDTO_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ClienteResponseDTO"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ClienteResponseDTO]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}